Moves regression tests to their own suite
