+++ test/acceptance/server_gettile.js
@@ -2,7 +2,6 @@
 require('../support/test_helper');
 
 var assert = require('../support/assert');
-var _ = require('underscore');
 var querystring = require('querystring');
 var fs = require('fs');
 var redis = require('redis');
@@ -292,150 +291,6 @@ suite('server_gettile', function() {
         done();
       });
 
-    });
-
-    // Test that you cannot write to the database from a tile request
-    //
-    // See http://github.com/CartoDB/Windshaft/issues/130
-    // Needs a fix on the mapnik side:
-    // https://github.com/mapnik/mapnik/pull/2143
-    //
-    // TODO: enable based on Mapnik version ?
-    //
-    test.skip("database access is read-only", function(done) {
-
-      step(
-        function doGet() {
-          var next = this;
-          assert.response(server, {
-              url: '/database/windshaft_test/table/test_table/0/0/0.png?sql=select+st_point(0,0)+as+the_geom,*+from+test_table_inserter(st_setsrid(st_point(0,0),4326),\'write\')',
-              method: 'GET'
-          }, {}, function(res, err) { next(err, res); });
-        },
-        function check(err, res) {
-          if ( err ) {
-              throw err;
-          }
-          assert.equal(res.statusCode, 400, res.statusCode + ': ' + ( res.statusCode !== 200 ? res.body : '..' ));
-          var parsed = JSON.parse(res.body);
-          assert.ok(parsed.error);
-          var msg = parsed.error;
-          assert.ok(msg.match(/read-only transaction/), msg);
-          return null;
-        },
-        function finish(err) {
-          assert.response(server, {
-              url: '/database/windshaft_test/table/test_table/style',
-              method: 'DELETE' },{}, function() { done(err); });
-        }
-      );
-
-    });
-
-    // See https://github.com/CartoDB/Windshaft/issues/167
-    test.skip("does not die on unexistent statsd host",  function(done) {
-      step(
-        function change_config() {
-          var CustomOptions = _.clone(ServerOptions);
-          CustomOptions.statsd = _.clone(CustomOptions.statsd);
-          CustomOptions.statsd.host = 'whoami.vizzuality.com';
-          CustomOptions.statsd.cacheDns = false;
-          server = new Windshaft.Server(CustomOptions);
-          server.setMaxListeners(0);
-          return null;
-        },
-        function do_get(err) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          var errors = [];
-          // We need multiple requests to make sure
-          // statsd_client eventually tries to send
-          // stats _and_ DNS lookup is given enough
-          // time (an interval is used later for that)
-          var numreq = 10;
-          var pending = numreq;
-          var completed = function(err) {
-            if ( err ) {
-                errors.push(err);
-            }
-            if ( ! --pending ) {
-              setTimeout(function() {
-              next(errors.length ? new Error(errors.join(',')) : null);
-              }, 10);
-              return;
-            }
-          };
-          for (var i=0; i<numreq; ++i) {
-            assert.response(server, {
-                url: '/database/windshaft_test/table/test_table/6/31/24.png',
-                method: 'GET'
-            },{}, function(res, err) { completed(err); });
-          }
-        },
-        function do_check(err) {
-          if ( err ) {
-              throw err;
-          }
-          // being alive is enough !
-          return null;
-        },
-        function finish(err) {
-          // reset server
-          server = new Windshaft.Server(ServerOptions);
-          done(err);
-        }
-      );
-    });
-
-    // See https://github.com/CartoDB/Windshaft/issues/173
-    test.skip("does not send db details in connection error response",  function(done) {
-      var base_key = 'map_style|windshaft_test|test_table';
-      step(
-        function change_config() {
-          var CustomOptions = _.clone(ServerOptions);
-          CustomOptions.grainstore = _.clone(CustomOptions.grainstore);
-          CustomOptions.grainstore.datasource = _.clone(CustomOptions.grainstore.datasource);
-          CustomOptions.grainstore.datasource.port = '666';
-          server = new Windshaft.Server(CustomOptions);
-          server.setMaxListeners(0);
-          return null;
-        },
-        function do_get(err) {
-          if ( err ) {
-              throw err;
-          }
-          var next = this;
-          assert.response(server, {
-              url: '/database/windshaft_test/table/test_table/6/31/24.png',
-              method: 'GET'
-          },{}, function(res, err) { next(err, res); });
-        },
-        function do_check(err, res) {
-          if ( err ) {
-              throw err;
-          }
-          // TODO: should be 500 !
-          assert.equal(res.statusCode, 400);
-          var parsed = JSON.parse(res.body);
-          assert.ok(parsed.error);
-          var msg = parsed.error;
-          assert.ok(msg.match(/connect/), msg);
-          assert.ok(!msg.match(/666/), msg);
-          return null;
-        },
-        function finish(err) {
-          // reset server
-          server = new Windshaft.Server(ServerOptions);
-          redis_client.del(base_key, function(e) {
-            if ( e ) {
-                console.error(e);
-            }
-            done(err);
-          });
-        }
-      );
     });
 
     test('high cpu regression with mapnik <2.3.x', function(done) {
